apps/ai-service/pyproject.toml
[project]
name = "chill-ai-service"
version = "0.0.1"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "openai>=1.50.0",
    "azure-identity>=1.19.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[tool.uv]
package = false
